feat: split chunk for pure string
