+++ src/memos/mem_reader/read_multi_modal/file_content_parser.py
@@ -16,7 +16,7 @@ from memos.memories.textual.item import (
 from memos.types.openai_chat_completion_types import File
 
 from .base import BaseMessageParser, _derive_key
-from .utils import get_parser, get_text_splitter
+from .utils import get_parser
 
 
 logger = get_logger(__name__)
@@ -107,33 +107,6 @@ class FileContentParser(BaseMessageParser):
                 ]
             else:
                 self.direct_markdown_hostnames = []
-
-    def _split_text(self, text: str) -> list[str]:
-        """
-        Split text into chunks using text splitter from utils.
-
-        Args:
-            text: Text to split
-
-        Returns:
-            List of text chunks
-        """
-        if not text or not text.strip():
-            return []
-
-        splitter = get_text_splitter()
-        if not splitter:
-            # If text splitter is not available, return text as single chunk
-            return [text] if text.strip() else []
-
-        try:
-            chunks = splitter.split_text(text)
-            logger.debug(f"[FileContentParser] Split text into {len(chunks)} chunks")
-            return chunks
-        except Exception as e:
-            logger.error(f"[FileContentParser] Error splitting text: {e}")
-            # Fallback to single chunk
-            return [text] if text.strip() else []
 
     def create_source(
         self,
